+++ dev/test-studio/components/navbar/ToolMenu.tsx
@@ -1,11 +1,12 @@
-import {Button, Flex} from '@sanity/ui'
+import {Button} from '@sanity/ui'
 import {ToolLink, type ToolMenuProps} from 'sanity'
+import {Flex} from 'ui5'
 
 export function ToolMenu(props: ToolMenuProps) {
   const {context, tools, closeSidebar} = props
 
   return (
-    <Flex gap={3} direction={context === 'sidebar' ? 'column' : 'row'}>
+    <Flex gap={3} flexDirection={context === 'sidebar' ? 'column' : 'row'}>
       {tools.map((tool) => (
         <ToolLink key={tool.name} name={tool.name}>
           {tool.title}