+++ packages/sanity/src/core/components/collapseMenu/CollapseMenu.tsx
@@ -1,4 +1,3 @@
-import {Flex} from '@sanity/ui'
 import difference from 'lodash-es/difference.js'
 import {
   Children,
@@ -11,6 +10,7 @@ import {
   useState,
   type RefAttributes,
 } from 'react'
+import {Flex, type GapProps, type MarginProps} from 'ui5'
 
 import {type MenuButtonProps} from '../../../ui-components/menuButton/MenuButton'
 import {Tooltip} from '../../../ui-components/tooltip/Tooltip'
@@ -26,7 +26,7 @@ export interface CollapseMenuProps {
   collapsed?: boolean
   collapseText?: boolean
   disableRestoreFocusOnClose?: boolean
-  gap?: number | number[]
+  gap?: GapProps['gap']
   menuButtonProps?: Omit<MenuButtonProps, 'id' | 'menu' | 'button'> & {
     id?: string
     button?: React.JSX.Element
@@ -237,15 +237,21 @@ export function AutoCollapseMenu(
 
   return (
     <Flex
-      align="center"
+      alignItems="center"
       data-ui="CollapseMenu"
       overflow="hidden"
-      sizing="border"
       ref={ref}
       {...rest}
       className={outerFlex}
     >
-      <Flex className={rootFlex} direction="column" flex={1} justify="center" ref={setRootEl}>
+      <Flex
+        className={rootFlex}
+        flexDirection="column"
+        flexBasis="0%"
+        flexGrow={1}
+        justifyContent="center"
+        ref={setRootEl}
+      >
         {/* The actual visible options */}
         <Flex className={rowFlex} gap={gap}>
           {pendingIntersections.length === 0 &&
@@ -288,7 +294,7 @@ export function AutoCollapseMenu(
 
       {/* Show the collapsed items that doesn't fit in a menu */}
       {overflowingCollapsedOptionElements.length > 0 && (
-        <Flex marginLeft={gap}>
+        <Flex marginLeft={gap as MarginProps['marginLeft']}>
           <CollapseOverflowMenu
             disableRestoreFocusOnClose={disableRestoreFocusOnClose}
             menuButton={menuButton}
@@ -304,7 +310,7 @@ export function AutoCollapseMenu(
 
 const RenderHidden = memo(function RenderHidden(props: {
   elements: React.JSX.Element[]
-  gap?: number | number[]
+  gap?: GapProps['gap']
   intersectionOptions: IntersectionObserverInit
   onIntersectionChange: (e: IntersectionObserverEntry, element: React.JSX.Element) => void
 }) {