+++ packages/sanity/src/core/scheduled-publishing/components/validation/ValidationInfo.tsx
@@ -6,11 +6,12 @@ import {
   type SchemaType,
   type ValidationMarker,
 } from '@sanity/types'
-import {type CardTone, Container, Stack} from '@sanity/ui'
+import {type CardTone, Container} from '@sanity/ui'
 import {Menu} from '@sanity/ui/menu'
 import * as PathUtils from '@sanity/util/paths'
 import {type ReactNode, useCallback, useId} from 'react'
 import {useRouter} from 'sanity/router'
+import {VStack} from 'ui5'
 
 import {Button} from '../../../../ui-components/button/Button'
 import {MenuButton} from '../../../../ui-components/menuButton/MenuButton'
@@ -72,14 +73,14 @@ export function ValidationInfo(props: ValidationProps) {
       menu={
         <Menu padding={1}>
           <Container width={0}>
-            <Stack gap={1}>
+            <VStack gap={1}>
               {menuHeader ?? null}
               <ValidationList
                 documentType={type as ObjectSchemaType}
                 validation={markers}
                 onFocus={onFocus}
               />
-            </Stack>
+            </VStack>
           </Container>
         </Menu>
       }